식단 분석, 로그인, 로그아웃, 식당 메뉴 관리 화면

+++ app/src/main/java/com/example/myapplication/service/DietService.kt
@@ -6,20 +6,18 @@ import java.time.LocalDate
 
 class DietService {
 
-    // 식단 추가
+    // 식단 추가 : 구현 X
     fun saveDiet(diet: Diet, callback: (Diet?) -> Unit) {
         // 식단 정보 저장
         callback(diet) // 저장 객체 반환
     }
-
-    // 식단 수정
+    // 식단 수정 : 구현 X
     fun updateDiet(diet: Diet, callback: (Diet?) -> Unit) {
         // 식단 정보 수정
         callback(diet)
     }
-
-    // 식단 삭제
-    fun deleteDiet(dietId: Int, callback: (Boolean) -> Unit) {
+    // 식단 삭제 : 구현 X
+    fun deleteDiet(dietId: Long, callback: (Boolean) -> Unit) {
         // 식단 정보 삭제
         callback(true)
     }
